+++ test/aws_request_test.dart
@@ -1,5 +1,5 @@
 import 'package:aws_request/aws_request.dart';
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
 
 void main() {
   test('tests getter / setter values', () {